+++ packages/shaker/src/plugins/shaker-plugin.ts
@@ -17,8 +17,9 @@ import {
   removeWithRelated,
   sideEffectImport,
   reference,
-  findParentForDelete,
+  findActionForNode,
   dereference,
+  mutate,
 } from '@linaria/utils';
 
 type Core = typeof core;
@@ -35,6 +36,10 @@ export interface IShakerMetadata {
 
 export interface IMetadata {
   __linariaShaker: IShakerMetadata;
+}
+
+interface NodeWithName {
+  name: string;
 }
 
 export const hasShakerMetadata = (
@@ -195,8 +200,14 @@ export default function shakerPlugin(
 
       if (!onlyExports.includes('*')) {
         const aliveExports = new Set<IExport | IReexport>();
+        const importNames = collected.imports.map(({ imported }) => imported);
+
         exports.forEach((exp) => {
           if (onlyExports.includes(exp.exported)) {
+            aliveExports.add(exp);
+          } else if (
+            importNames.includes((exp.local.node as NodeWithName).name || '')
+          ) {
             aliveExports.add(exp);
           }
         });
@@ -257,7 +268,8 @@ export default function shakerPlugin(
           // eslint-disable-next-line no-restricted-syntax
           for (const path of forDeleting) {
             const binding = getBindingForExport(path);
-            const parent = findParentForDelete(path);
+            const action = findActionForNode(path);
+            const parent = action?.[1];
             const outerReferences = (binding?.referencePaths || []).filter(
               (ref) => ref !== parent && !parent?.isAncestor(ref)
             );
@@ -271,7 +283,20 @@ export default function shakerPlugin(
               !deleted.has(path) &&
               (!binding || outerReferences.length === 0)
             ) {
-              removeWithRelated([parent ?? path]);
+              if (action) {
+                mutate(action[1], (p) => {
+                  if (isRemoved(p)) return;
+
+                  if (action[0] === 'remove') {
+                    p.remove();
+                  } else if (action[0] === 'replace') {
+                    p.replaceWith(action[2]);
+                  }
+                });
+              } else {
+                removeWithRelated([path]);
+              }
+
               deleted.add(path);
               changed = true;
             }