+++ packages/react/CHANGELOG.md
@@ -1,5 +1,20 @@
 # Change Log
 
+## 4.3.0
+
+### Minor Changes
+
+- 63f56d47: Do not filter properties if an unknown component is passed to `styled`. Fixes support of custom elements #968
+
+### Patch Changes
+
+- c26d4667: force interop check to fix @emotion/is-prop-valid esm import
+- Updated dependencies [63f56d47]
+- Updated dependencies [963508a2]
+  - @linaria/tags@4.2.0
+  - @linaria/utils@4.2.4
+  - @linaria/core@4.2.2
+
 ## 4.2.1
 
 ### Patch Changes