+++ packages/utils/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## 4.2.4
+
+### Patch Changes
+
+- 963508a2: Shaker shouldn't remove parameters of functions if they aren't used.
+
 ## 4.2.3
 
 ### Patch Changes